Phase 1: extract features/theme.js
setupTheme/toggleTheme/window.setTheme + updateThemeIcon move out. The mermaid
re-render (render core) is supplied via configureTheme({ reRenderDiagrams }),
wired lazily in init so it resolves the current function at call time.
main.js ~1,808 -> ~1,776. Build + lint clean; 297 tests pass.
